pyproject.toml
[project]
name = "synapse_delete_room_rest_api"
description = "This module expose and endpoint for room admins (or room members with highest power level) to kick everyone out and leave the room."
readme = "README.md"

version = "0.0.1"


requires-python = ">=3.9"

classifiers = [
  "License :: OSI Approved :: Apache Software License"
]

dependencies = [
  "attrs"
]

[project.optional-dependencies]
dev = [
  # for tests
  "matrix-synapse",
  "tox",
  "twisted == 24.7.0",
  "aiounittest",
  "psycopg2",
  "testing.postgresql",
  # for type checking
  "mypy == 1.6.1",
  # for linting
  "black == 23.10.0",
  "ruff == 0.1.1",
]

[build-system]
requires = [
  "setuptools",
  "wheel",
  "setuptools_scm",
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

[tool.mypy]
files = ["synapse_delete_room_rest_api", "tests"]
disable_error_code = ["import-untyped", "no-untyped-call", "attr-defined"]
strict = false 

[tool.ruff]
line-length = 88

# See https://docs.astral.sh/ruff/rules/#error-e
# for error codes. The ones we ignore are:
#  E501: Line too long (black enforces this for us)
#  E731: do not assign a lambda expression, use a def
#
# flake8-bugbear compatible checks. Its error codes are described at
# https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
#  B023: Functions defined inside a loop must not use variables redefined in the loop
ignore = [
    "B023",
    "E501",
    "E731",
]
select = [
    # pycodestyle
    "E",
    "W",
    # pyflakes
    "F",
    # flake8-bugbear
    "B0",
    # flake8-comprehensions
    "C4",
    # flake8-2020
    "YTT",
    # flake8-slots
    "SLOT",
    # flake8-debugger
    "T10",
    # flake8-pie
    "PIE",
    # flake8-executable
    "EXE",
    # isort
    "I",
]
